Split dataset.py into separate module and test files
- Extract PercolationDataset and Node classes to percolation_dataset.py
- Move unit tests to test_percolation_dataset.py
- Add proper import and __main__ guard for test execution
